hotplug: add new x11hotplug backend
For systems without udev or hald+dbus, xorg hotplugging is not possible.
This patch adds another hotplug backend, that does not depend on other
infrastructure as udev etc.

The backend comes with a program, x11hotplug, that allows to insert
devices to the xorg server. The idea is that the administrator
can script his or her devices to be added, with the appropriate parameters.
Therefore, less xorg.conf magic is required because parameters can
be provided by the caller.
x11hotplug may even be called from udev events.

Signed-off-by: Kurt Van Dijck <kurt@vandijck-laurijssen.be>